Add options argument to several methods
